+++ src/test/java/com/stripe/StripeTest.java
@@ -9,15 +9,12 @@ import com.stripe.exception.InvalidRequestException;
 import com.stripe.exception.StripeException;
 import com.stripe.model.Account;
 import com.stripe.model.AccountCollection;
-import com.stripe.model.Address;
 import com.stripe.model.AlipayAccount;
 import com.stripe.model.ApplePayDomain;
-import com.stripe.model.ApplePayDomainCollection;
 import com.stripe.model.ApplicationFee;
 import com.stripe.model.Balance;
 import com.stripe.model.BalanceTransaction;
 import com.stripe.model.BalanceTransactionCollection;
-import com.stripe.model.BankAccount;
 import com.stripe.model.BitcoinReceiver;
 import com.stripe.model.BitcoinTransaction;
 import com.stripe.model.Card;
@@ -30,7 +27,6 @@ import com.stripe.model.Customer;
 import com.stripe.model.CustomerSubscriptionCollection;
 import com.stripe.model.SubscriptionCollection;
 import com.stripe.model.DeletedApplePayDomain;
-import com.stripe.model.DeletedBankAccount;
 import com.stripe.model.DeletedBitcoinReceiver;
 import com.stripe.model.DeletedCard;
 import com.stripe.model.DeletedCoupon;
@@ -41,13 +37,9 @@ import com.stripe.model.DeletedProduct;
 import com.stripe.model.DeletedRecipient;
 import com.stripe.model.DeletedSKU;
 import com.stripe.model.DeletedSubscriptionItem;
-import com.stripe.model.Dispute;
 import com.stripe.model.Event;
-import com.stripe.model.EvidenceDetails;
-import com.stripe.model.EvidenceSubObject;
 import com.stripe.model.ExternalAccount;
 import com.stripe.model.ExternalAccountCollection;
-import com.stripe.model.FileUpload;
 import com.stripe.model.FraudDetails;
 import com.stripe.model.Invoice;
 import com.stripe.model.InvoiceItem;
@@ -62,7 +54,6 @@ import com.stripe.model.Recipient;
 import com.stripe.model.Refund;
 import com.stripe.model.Source;
 import com.stripe.model.SKU;
-import com.stripe.model.ShippingDetails;
 import com.stripe.model.Subscription;
 import com.stripe.model.SubscriptionItem;
 import com.stripe.model.SubscriptionItemCollection;
@@ -78,11 +69,6 @@ import junit.framework.Assert;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
-
-import java.io.File;
-
-import java.net.URISyntaxException;
-import java.net.URL;
 
 import java.util.Calendar;
 import java.util.Date;
@@ -467,263 +453,6 @@ public class StripeTest {
 		listParams.put("count", 1);
 		List<Charge> charges = Charge.all(listParams).getData();
 		assertEquals(charges.size(), 1);
-	}
-
-	// Dispute Tests:
-	@Test
-	public void testDisputedCharge() throws StripeException, InterruptedException {
-		int chargeValueCents = 100;
-		Charge disputedCharge = createDisputedCharge(chargeValueCents);
-		Dispute dispute = disputedCharge.getDispute();
-		assertNotNull(dispute);
-		assertFalse(dispute.getIsChargeRefundable());
-		assertEquals(1, dispute.getBalanceTransactions().size());
-		assertEquals(-chargeValueCents, dispute.getBalanceTransactions().get(0).getAmount().intValue());
-	}
-
-	@Test
-	public void testRetrieveDispute() throws StripeException, InterruptedException {
-		int chargeValueCents = 100;
-		Charge disputedCharge = createDisputedCharge(chargeValueCents);
-		Dispute dispute = disputedCharge.getDispute();
-		Dispute retrievedDispute = Dispute.retrieve(dispute.getId());
-		assertEquals(dispute.getId(), retrievedDispute.getId());
-	}
-
-	@Test
-	public void testUpdateDispute() throws StripeException, InterruptedException {
-		int chargeValueCents = 100;
-		Charge disputedCharge = createDisputedCharge(chargeValueCents);
-
-		Dispute initialDispute = disputedCharge.getDispute();
-		EvidenceSubObject emptyEvidence = new EvidenceSubObject();
-		assertEquals(emptyEvidence, initialDispute.getEvidenceSubObject());
-		assertEquals(new HashMap<String, String>(), initialDispute.getMetadata());
-
-		Map<String, Object> params = new HashMap<String, Object>();
-		Map<String, Object> evidence = new HashMap<String, Object>();
-		Map<String, String> metadata = new HashMap<String, String>();
-
-		evidence.put("product_description", "my productDescription");
-		evidence.put("customer_name", "my customerName");
-		evidence.put("uncategorized_text", "my uncategorizedText");
-
-		metadata.put("some_info", "about the dispute");
-		metadata.put("a_little_more", "12345");
-
-		params.put("evidence", evidence);
-		params.put("metadata", metadata);
-
-		Dispute updatedDispute = initialDispute.update(params);
-		assertNotNull(updatedDispute);
-
-		EvidenceSubObject evidenceSubObject = updatedDispute.getEvidenceSubObject();
-		assertEquals(evidence.get("product_description"), evidenceSubObject.getProductDescription());
-		assertEquals(evidence.get("customer_name"), evidenceSubObject.getCustomerName());
-		assertEquals(evidence.get("uncategorized_text"), evidenceSubObject.getUncategorizedText());
-
-		// Ensure this didn't get stored in the deprecated evidence field.
-		assertNull(updatedDispute.getEvidence());
-
-		Map<String, String> disputeMetadata = updatedDispute.getMetadata();
-		assertNotNull(disputeMetadata);
-		assertEquals(metadata, disputeMetadata);
-	}
-
-	@Test
-	public void testCloseDispute() throws StripeException, InterruptedException {
-		int chargeValueCents = 100;
-		Charge disputedCharge = createDisputedCharge(chargeValueCents);
-		Dispute dispute = disputedCharge.getDispute();
-		assertEquals("needs_response", dispute.getStatus());
-
-		Dispute closedDispute = dispute.close();
-		assertEquals(dispute.getId(), closedDispute.getId());
-		assertEquals("lost", closedDispute.getStatus());
-	}
-
-	// DisputeCollection Tests:
-	@Test
-	public void testDisputeList() throws StripeException, InterruptedException {
-		Map<String, Object> listParams = new HashMap<String, Object>();
-		listParams.put("count", 3);
-		List<Dispute> disputes = Dispute.all(listParams).getData();
-
-		assertEquals(3, disputes.size());
-	}
-
-	// FraudDetails Test:
-	@Test
-	public void testFraudDetails() throws StripeException, InterruptedException {
-		Charge charge = Charge.create(defaultChargeParams);
-		FraudDetails expected = new FraudDetails();
-		assertEquals(expected, charge.getFraudDetails());
-		Charge refundedCharge = charge.refund();
-		assertEquals(expected, refundedCharge.getFraudDetails());
-		Charge updatedCharge = charge.update(ImmutableMap.<String, Object>of("fraud_details", ImmutableMap.of("user_report", "fraudulent")));
-		FraudDetails expectedReported = new FraudDetails();
-		expectedReported.setUserReport("fraudulent");
-		assertEquals(expectedReported, updatedCharge.getFraudDetails());
-
-		Charge nowSafe = updatedCharge.markSafe(null);
-		expectedReported.setUserReport("safe");
-		assertEquals(expectedReported, nowSafe.getFraudDetails());
-
-		Charge nowFraudulent = nowSafe.markFraudulent(null);
-		expectedReported.setUserReport("fraudulent");
-		assertEquals(expectedReported, nowFraudulent.getFraudDetails());
-	}
-
-	// FileUpload Tests:
-	@Test
-	public void testCreateFileUpload() throws StripeException,
-			InterruptedException, URISyntaxException {
-		URL url = getClass().getResource("minimal.pdf");
-		File file = new File(url.getPath());
-		Map<String, Object> fileUploadParams = new HashMap<String, Object>();
-		fileUploadParams.put("purpose", "dispute_evidence");
-		fileUploadParams.put("file", file);
-
-		FileUpload fileUpload = FileUpload.create(fileUploadParams);
-		assertEquals(file.length(), fileUpload.getSize().longValue());
-		assertEquals("pdf", fileUpload.getType());
-
-		FileUpload retrievedUpload = FileUpload.retrieve(fileUpload.getId());
-		assertEquals(fileUpload.getId(), retrievedUpload.getId());
-		assertEquals(file.length(), retrievedUpload.getSize().longValue());
-		assertEquals("pdf", retrievedUpload.getType());
-	}
-
-	@Test
-	public void testFileUploadList() throws StripeException {
-		Map<String, Object> listParams = new HashMap<String, Object>();
-		listParams.put("limit", 1);
-		List<FileUpload> uploads = FileUpload.all(listParams).getData();
-		assertEquals(uploads.size(), 1);
-	}
-
-	@Test
-	public void testSubmitOldStyleEvidence() throws StripeException, InterruptedException {
-		int chargeValueCents = 100;
-		Stripe.apiVersion = "2014-11-20";
-		Charge disputedCharge = createDisputedCharge(chargeValueCents);
-
-		String myEvidence = "Here's evidence showing this charge is legitimate.";
-		Dispute initialDispute = disputedCharge.getDispute();
-		assertNull(initialDispute.getEvidence());
-		assertNull(initialDispute.getEvidenceSubObject());
-		Map<String, Object> disputeParams = ImmutableMap.<String, Object>of("evidence", myEvidence);
-
-		Dispute updatedDispute = disputedCharge.updateDispute(disputeParams);
-		assertNotNull(updatedDispute);
-		assertEquals(myEvidence, updatedDispute.getEvidence());
-		assertNull(updatedDispute.getEvidenceSubObject());
-	}
-
-	@Test
-	public void testSubmitEvidence() throws StripeException, InterruptedException {
-		int chargeValueCents = 100;
-		Charge disputedCharge = createDisputedCharge(chargeValueCents);
-
-		Dispute initialDispute = disputedCharge.getDispute();
-		assertNull(initialDispute.getEvidence());
-		EvidenceSubObject emptyEvidence = new EvidenceSubObject();
-		assertEquals(emptyEvidence, initialDispute.getEvidenceSubObject());
-		assertEquals(0, initialDispute.getEvidenceDetails().getSubmissionCount().intValue());
-
-		Map<String, Object> evidenceHashParams = new HashMap<String, Object>();
-		// TODO: assert on all param types
-		evidenceHashParams.put("product_description", "my productDescription");
-		evidenceHashParams.put("customer_name", "my customerName");
-		evidenceHashParams.put("uncategorized_text", "my uncategorizedText");
-		Map<String, Object> providedEvidenceParams = ImmutableMap.<String, Object>of("evidence", evidenceHashParams);
-
-		Dispute updatedDispute = disputedCharge.updateDispute(providedEvidenceParams);
-		assertNotNull(updatedDispute);
-		EvidenceSubObject evidenceSubObject = updatedDispute.getEvidenceSubObject();
-		assertNotSame(emptyEvidence, evidenceSubObject);
-		assertEquals(1, updatedDispute.getEvidenceDetails().getSubmissionCount().intValue());
-		assertNull(updatedDispute.getEvidence());
-
-		assertEquals("my productDescription", evidenceSubObject.getProductDescription());
-		assertEquals("my customerName", evidenceSubObject.getCustomerName());
-		assertEquals("my uncategorizedText", evidenceSubObject.getUncategorizedText());
-
-		EvidenceDetails evidenceDetails = updatedDispute.getEvidenceDetails();
-		assertNotNull(evidenceDetails);
-		assertEquals(1, evidenceDetails.getSubmissionCount().intValue());
-	}
-
-	private Charge createDisputedCharge(int chargeValueCents) throws AuthenticationException, InvalidRequestException, APIConnectionException, CardException, APIException, InterruptedException {
-		Map<String, Object> chargeParams = new HashMap<String, Object>();
-		chargeParams.putAll(defaultChargeParams);
-		chargeParams.put("amount", chargeValueCents);
-		Map<String, Object> testModeDisputeCardParams = new HashMap<String, Object>();
-		testModeDisputeCardParams.put("number", "4000000000000259");
-		testModeDisputeCardParams.put("exp_month", 12);
-		testModeDisputeCardParams.put("exp_year", getYear());
-		chargeParams.put("card", testModeDisputeCardParams);
-		Charge charge = Charge.create(chargeParams);
-
-		// This test relies on the server asynchronously marking the charge as disputed.
-		// TODO: find a more reliable way to do this instead of sleeping
-		Thread.sleep(10000);
-		return Charge.retrieve(charge.getId());
-	}
-
-	// Plan Tests:
-	@Test
-	public void testPlanCreate() throws StripeException {
-		Plan plan = Plan.create(getUniquePlanParams());
-		assertEquals(plan.getInterval(), "month");
-		assertEquals(plan.getIntervalCount(), (Integer) 2);
-	}
-
-	@Test
-	public void testPlanCreateWithStatementDescriptor() throws StripeException {
-		Map<String, Object> planParamsWithStatementDescriptor = getUniquePlanParams();
-		planParamsWithStatementDescriptor.put("statement_descriptor", "Stripe");
-		Plan plan = Plan.create(planParamsWithStatementDescriptor);
-		assertEquals(plan.getStatementDescriptor(), "Stripe");
-	}
-
-	@Test
-	public void testPlanUpdate() throws StripeException {
-		Plan createdPlan = Plan.create(getUniquePlanParams());
-		Map<String, Object> updateParams = new HashMap<String, Object>();
-		updateParams.put("name", "Updated Plan Name");
-		Plan updatedplan = createdPlan.update(updateParams);
-		assertEquals(updatedplan.getName(), "Updated Plan Name");
-	}
-
-	@Test
-	public void testPlanRetrieve() throws StripeException {
-		Plan createdPlan = Plan.create(getUniquePlanParams());
-		Plan retrievedPlan = Plan.retrieve(createdPlan.getId());
-		assertEquals(createdPlan.getId(), retrievedPlan.getId());
-	}
-
-	@Test
-	public void testPlanList() throws StripeException {
-		Map<String, Object> listParams = new HashMap<String, Object>();
-		listParams.put("count", 1);
-		List<Plan> Plans = Plan.all(listParams).getData();
-		assertEquals(Plans.size(), 1);
-	}
-
-	@Test
-	public void testPlanDelete() throws StripeException {
-		Plan createdPlan = Plan.create(getUniquePlanParams());
-		DeletedPlan deletedPlan = createdPlan.delete();
-		assertTrue(deletedPlan.getDeleted());
-		assertEquals(deletedPlan.getId(), createdPlan.getId());
-	}
-
-	@Test
-	public void testCustomerCreateWithPlan() throws StripeException {
-		Plan plan = Plan.create(getUniquePlanParams());
-		Customer customer = createDefaultCustomerWithPlan(plan);
-		assertEquals(customer.getSubscriptions().getData().get(0).getPlan().getId(), plan.getId());
 	}
 
 	// Subscription Tests:
